+++ src/main/java/lotto/domain/lottonumber/LottoBalls.java
@@ -1,14 +1,12 @@
-package lotto;
+package lotto.domain.lottonumber;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import lotto.domain.LottoNumber;
 
 public class LottoBalls {
