+++ src/test/java/lotto/domain/LottoRanksTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.domain.lottorank.LottoRank;
+import lotto.domain.lottorank.LottoRanks;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
